+++ eoreader/products/optical/l1_product.py
@@ -15,7 +15,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """ Landsat-1 products """
-from eoreader.products.optical.landsat_product import LandsatProduct
+from eoreader.products import LandsatProduct
 
 
 class L1Product(LandsatProduct):