+++ eoreader/products/optical/s3_slstr_product.py
@@ -39,17 +39,12 @@ from sertit.rasters import MAX_CORES, XDS_TYPE
 from sertit.vectors import WGS84
 
 from eoreader import cache, utils
-from eoreader.bands.alias import ALL_CLOUDS, CIRRUS, CLOUDS, RAW_CLOUDS, to_str
-from eoreader.bands.bands import BandNames
-from eoreader.bands.bands import OpticalBandNames as obn
+from eoreader.bands import ALL_CLOUDS, CIRRUS, CLOUDS, RAW_CLOUDS, BandNames
+from eoreader.bands import OpticalBandNames as obn
+from eoreader.bands import to_str
 from eoreader.exceptions import InvalidProductError, InvalidTypeError
 from eoreader.keywords import SLSTR_RAD_ADJUST, SLSTR_STRIPE, SLSTR_VIEW
-from eoreader.products.optical.s3_product import (
-    S3DataType,
-    S3Instrument,
-    S3Product,
-    S3ProductType,
-)
+from eoreader.products import S3DataType, S3Instrument, S3Product, S3ProductType
 from eoreader.reader import Platform
 from eoreader.utils import EOREADER_NAME