+++ eoreader/products/sar/cosmo_product.py
@@ -36,7 +36,7 @@ from shapely.geometry import Polygon, box
 
 from eoreader import cache, cached_property
 from eoreader.exceptions import InvalidProductError
-from eoreader.products.sar.sar_product import SarProduct, SarProductType
+from eoreader.products import SarProduct, SarProductType
 from eoreader.utils import DATETIME_FMT, EOREADER_NAME
 
 LOGGER = logging.getLogger(EOREADER_NAME)