+++ eoreader/products/sar/csg_product.py
@@ -31,7 +31,7 @@ from sertit.rasters import XDS_TYPE
 
 from eoreader.bands.bands import BandNames
 from eoreader.exceptions import InvalidProductError
-from eoreader.products.sar.cosmo_product import CosmoProduct, CosmoProductType
+from eoreader.products import CosmoProduct, CosmoProductType
 from eoreader.utils import EOREADER_NAME
 
 LOGGER = logging.getLogger(EOREADER_NAME)