+++ eoreader/products/sar/iceye_product.py
@@ -34,7 +34,7 @@ from sertit.misc import ListEnum
 from eoreader import cache, cached_property
 from eoreader.bands.bands import SarBandNames as sbn
 from eoreader.exceptions import InvalidProductError, InvalidTypeError
-from eoreader.products.sar.sar_product import SarProduct, SarProductType
+from eoreader.products import SarProduct, SarProductType
 from eoreader.utils import DATETIME_FMT, EOREADER_NAME
 
 LOGGER = logging.getLogger(EOREADER_NAME)