+++ eoreader/products/sar/rcm_product.py
@@ -34,7 +34,7 @@ from sertit.vectors import WGS84
 
 from eoreader import cache, cached_property
 from eoreader.exceptions import InvalidProductError, InvalidTypeError
-from eoreader.products.sar.sar_product import SarProduct, SarProductType
+from eoreader.products import SarProduct, SarProductType
 from eoreader.utils import DATETIME_FMT, EOREADER_NAME
 
 LOGGER = logging.getLogger(EOREADER_NAME)