+++ eoreader/products/sar/tsx_product.py
@@ -32,7 +32,7 @@ from sertit.misc import ListEnum
 
 from eoreader import cache, cached_property
 from eoreader.exceptions import InvalidProductError, InvalidTypeError
-from eoreader.products.sar.sar_product import SarProduct, SarProductType
+from eoreader.products import SarProduct, SarProductType
 from eoreader.reader import Platform
 from eoreader.utils import DATETIME_FMT, EOREADER_NAME